refactor(maestro): split ide/diagnostics/corsa into cohesive modules
Pure structural refactor: move the ~1072-line corsa.rs into a
corsa/ subdir with cohesive submodules behind the unchanged
`mod corsa;` declaration in diagnostics.rs.

- collect.rs:     collect_corsa_diagnostics + sibling overlay
- virtual_ts.rs:  generate_virtual_ts / _for_art + helpers
- mapping.rs:     diagnostic range <-> source offset translation
- message.rs:     Vue-flavored message rewriting (+ hint tests)
- tests.rs:       relocated #[cfg(test)] mod tests
- mod.rs:         module declarations

Behavior, public API, and output are identical; only code moved
and module visibility adjusted (pub(super) preserved as
pub(in crate::ide::diagnostics) where the parent module needs it).
